마스터 병합 오류 해결

+++ src/main/java/com/example/demo/database/Entity/StudentEntity.java
@@ -28,7 +28,7 @@ public class StudentEntity implements Serializable {
     private String email;
     private String birth;
     private String address;
-    private Integer lecture;
+    private Long lecture; // type Long 이여야함
     private String gender;
     private String regdate;
     private String lecname;
